Разбил все по файлам, добавил обработку JSON
